split common/utils into server and client
