+++ apps/dialtone-documentation/docs/.vuepress/theme/client.js
@@ -5,7 +5,7 @@ import customEmojis from '@data/custom-emoji.json';
 import 'overlayscrollbars/overlayscrollbars.css';
 import { OverlayScrollbars, ClickScrollPlugin } from 'overlayscrollbars';
 import { onBeforeMount, provide, ref } from 'vue';
-import { flushPromises } from '@workspaceRoot/common/utils';
+import { flushPromises } from '@workspaceRoot/common/utils/client.mjs';
 
 // CSS
 import '@dialpad/dialtone-css/lib/dist/dialtone.css';
