+++ common/utils/client.mjs
@@ -1,4 +1,7 @@
-import { readdirSync } from 'fs';
+/**
+ * Contains utilities that will work on a browser,
+ * don't import anything related to node or that is not SSR friendly.
+ */
 
 let UNIQUE_ID_COUNTER = 0;
 export function getUniqueString (prefix = 'dt') {
@@ -36,34 +39,9 @@ export function PascalCaseToKebabCase (string) {
     .toLowerCase();
 }
 
-/**
- * Scans recursively through the provided path
- * and gets the valid contained Vue components and recipes.
- * @param {PathLike} folder
- * @returns {PathLike[]}
- */
-export function getValidFileList (folder) {
-  const parentFolderName = folder.split('/').pop();
-  const excludedFolderNamesRegex = /(extensions|modules|decorators)$/;
-  const validFileNamesRegex = new RegExp(`^${parentFolderName}\\.vue$`);
-
-  return readdirSync(folder, { withFileTypes: true })
-    .filter((item) => (
-      (item.isDirectory() && !excludedFolderNamesRegex.test(item.name)) ||
-      validFileNamesRegex.test(item.name)
-    ))
-  .reduce((files, item) => {
-    if (item.isDirectory()) files = [...files, ...getValidFileList(`${folder}/${item.name}`)]
-    else files.push(`${folder}/${item.name}`);
-
-    return files;
-  }, []);
-}
-
 export default {
   getUniqueString,
   kebabCaseToPascalCase,
   flushPromises,
-  PascalCaseToKebabCase,
-  getValidFileList
+  PascalCaseToKebabCase
 };
